refactor: Extracted model
refs: BLAIS5-3143
