+++ models/questionnaire_case_task_model.py
@@ -1,0 +1,15 @@
+import json
+
+from dataclasses import asdict, dataclass
+from typing import Dict
+
+
+@dataclass
+class QuestionnaireCaseTaskModel:
+    questionnaire: str
+
+    def as_dict(self) -> Dict:
+        return asdict(self)
+
+    def json(self) -> str:
+        return json.dumps(self.as_dict())
